% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGoogleTiles.R
\name{addGoogleTiles}
\alias{addGoogleTiles}
\title{Graphics elements and layers}
\usage{
addGoogleTiles(map, attribution = NULL, layerId = NULL, group = NULL,
  options = tileOptions(), type = "satellite")
}
\arguments{
\item{map}{a leaflet map object}

\item{attribution}{the attribution text of the tile layer (HTML)}

\item{layerId}{the identification number/string of the layer created}

\item{group}{the group identification number/string}

\item{options}{a list of extra options for tile layers, popups, paths
(circles, rectangles, polygons, ...), or other map elements}

\item{type}{map type (satellite, roadmap, hybrid)}
}
\value{
the new \code{map} object
}
\description{
Add graphics elements and layers to the map widget.
}
\references{
The Leaflet API documentation:
  \url{http://leafletjs.com/reference.html}
  \url{https://gitlab.com/IvanSanchez/Leaflet.GridLayer.GoogleMutant}
}
