% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGoogleTiles.R
\name{addGTiles}
\alias{addGTiles}
\title{Graphics elements and layers :}
\usage{
addGTiles(map)
}
\arguments{
\item{map}{the leaflet map}
}
\value{
the leaflet google layer map
}
\description{
Add the google layer to the map widget (without shiny).
}
