% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGoogleTiles.R
\name{initPlugin}
\alias{initPlugin}
\title{User functions :}
\usage{
initPlugin(map, apikey)
}
\arguments{
\item{map}{the leaflet map}

\item{apikey}{the api-key you want to use for your application (you need to use a valid key)}
}
\value{
the option to use the google base tile on leaflet without shiny
}
\description{
Use the package without shiny (this function works with the other function addGTiles)
}
