% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGoogleTiles.R
\name{polygonBoundaries}
\alias{polygonBoundaries}
\title{Polygon functions :}
\usage{
polygonBoundaries(liste)
}
\arguments{
\item{liste}{a list of the points of the polygon}
}
\value{
a vector composed by the limits of the square which wrap the polygon
}
\description{
Find the boundarie of a polygon, and the middle of it, given the list of their composite points.
}
