% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGoogleTiles.R
\name{setZoomBound}
\alias{setZoomBound}
\title{Zoom functions :}
\usage{
setZoomBound(xdist, ydist)
}
\arguments{
\item{xdist}{the horizontal distance of the map displayed in the widget, it's in degree}

\item{ydist}{the vertical distance of the map displayed in the widget, it's in degree}
}
\value{
a vector composed by the optimal zoom level and its dependencies
}
\description{
Set the optimal zoom for your raster using the boundaries of the raster.
}
\references{
The raster package:
  \url{https://cran.r-project.org/web/packages/raster/raster.pdf}
}
