% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGoogleTiles.R
\name{useLeafletGoogle}
\alias{useLeafletGoogle}
\title{User functions :}
\usage{
useLeafletGoogle()
}
\value{
a tags$head that will be usefull to load the references js/css
}
\description{
Make the package more user friendly (you need to use this function into the ui part of your shinny Application)
}
